DevC++ in 32 bit version defaults to 64 bit pointers for whatever reason i know its broken, that is the point, to make this crap immune to most broken compilers ever statically linked as exe to Kop32 works, compiles, dll compiles but wont link (Dllentry point exist, but main no, so wont link) in stasm module, my tweak to shared global (but protected atomic) variables to exchange data in/out of asm() for whatever reason wasn't recognized MinGW was eating "_" in name, goal is to compile agains very same old compiler over and over in variety of enviroments never use (if possible) C++ 11 there is no point for this library except old systems no C++ 11 support anyway if there is any point except me learning

+++ compression/stk_compression_API.cpp
@@ -22,8 +22,8 @@ void __stdcall stk::compression::test(void)
 __DEBUG_CALLED("")
 #endif
 	register int32_t o, i = 0;
-	char *src = stk::cstr::alloc(4096);
-	char *dst = stk::cstr::alloc(4096);
+	char *src = stk::cstr::alloc((const size_t)4096);
+	char *dst = stk::cstr::alloc((const size_t)4096);
 if (src==NULL || dst==NULL)
 return;
 	stk::con::prints("Adres src: %d\n",(int8_t*)src);
